requirements.txt
aioredis==1.2.0
asgiref==3.2.1
asn1crypto==0.24.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==2.2.3
djangorestframework==3.10.2
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.19
incremental==17.5.0
more-itertools==7.2.0
msgpack==0.6.1
packaging==19.1
pluggy==0.12.0
py==1.8.0
pycparser==2.19
PyHamcrest==1.9.0
pyparsing==2.4.2
pytest==5.0.1
pytest-asyncio==0.10.0
pytest-django==3.5.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
wcwidth==0.1.7
zipp==0.5.2
zope.interface==4.6.0
